move auth check on protected routes to auth middleware
